Code clean up, optimizations, and documentation
