+++ Assets/Scripts/Inventory/InventoryUIEntry.cs.meta
