--- Assets/Scripts/NewInventory/ItemEvent.cs
@@ -1,6 +1,0 @@
-using UnityEngine.Events;
-
-public class ItemEvent : UnityEvent
-{
-	public Item item;
-}
