--- Assets/Scripts/NewInventory/ItemEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b7bd30a7bb15c44c8ae56838f0d9fa6c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
