+++ Assets/Scripts/OldInventory.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 02411f24786774dfe9d019c8a6d928ea
+guid: fb94a734f00152043ac88aded747ab70
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
