+++ Assets/Scripts/OldInventory/ExampleItem.asset
