+++ Assets/Scripts/OldInventory/ExampleItem.asset.meta
