+++ Assets/Scripts/OldInventory/OldInventorySlot.cs.meta
