+++ Assets/Scripts/OldInventory/OldItemData.cs.meta
