+++ Assets/Scripts/OldInventory/OldItemPickup.cs.meta
